fix MSCV warnings

+++ kahypar/partition/preprocessing/policies/min_hash_policy.h
@@ -98,7 +98,7 @@ class MinHashPolicy {
     return _dim;
   }
 
-  uint32_t getHashNum() const {
+  size_t getHashNum() const {
     return _hash_func_vector.getHashNum();
   }
